BSP-333: refactoring
